[PAY-2997] Consolidate hashid usage
