+++ packages/common/src/utils/hashIds.ts
@@ -1,36 +1,39 @@
-import Hashids from 'hashids'
+import { decodeHashId, encodeHashId } from '@audius/sdk'
+import { z } from 'zod'
 
-import { Nullable } from './typeUtils'
+export const HashId = z.string().transform<number>((data: string, context) => {
+  const id = decodeHashId(data)
+  if (id === null) {
+    context.addIssue({
+      code: z.ZodIssueCode.custom,
+      message: 'Hash id is invalid'
+    })
 
-const HASH_SALT = 'azowernasdfoia'
-const MIN_LENGTH = 5
-const hashids = new Hashids(HASH_SALT, MIN_LENGTH)
-
-/** Decodes a string id into an int. Returns null if an invalid ID. */
-export const decodeHashId = (id?: string): Nullable<number> => {
-  if (!id) return null
-  try {
-    const ids = hashids.decode(id)
-    if (!ids.length) return null
-    const num = Number(ids[0])
-    if (isNaN(num)) return null
-    return num
-  } catch (e) {
-    console.error(`Failed to decode ${id}`, e)
-    return null
+    return z.NEVER
   }
-}
+  return id
+})
 
-export function encodeHashId(id: null): null
-export function encodeHashId(id: number): string
-export function encodeHashId(id: Nullable<number>): Nullable<string>
-export function encodeHashId(id: Nullable<number>): Nullable<string> {
-  try {
-    if (id === null) return null
-    const encodedId = hashids.encode(id)
-    return encodedId
-  } catch (e) {
-    console.error(`Failed to encode ${id}`, e)
-    return null
+/** Parses a `Nullable<ID>` safely for use with SDK functions which acccept an optional
+   id parameter.
+*/
+export const OptionalId = z
+  .number()
+  .nullable()
+  .optional()
+  .transform<string | undefined>((data: number | null | undefined) =>
+    data ? (encodeHashId(data) ?? undefined) : undefined
+  )
+
+export const Id = z.number().transform<string>((data: number, context) => {
+  const id = encodeHashId(data)
+  if (id === null) {
+    context.addIssue({
+      code: z.ZodIssueCode.custom,
+      message: 'Hash id is invalid'
+    })
+
+    return z.NEVER
   }
-}
+  return id
+})
